3_python_service/consumer_health_check_service/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
Werkzeug==3.0.3
zipp==3.18.1
